Only delete when there are files to delete

+++ examples/postgresql/update_samples.R
@@ -36,7 +36,7 @@ downloadSample = function(tab){
 
 by(allTables, 1:nrow(allTables), function(row){
   tab = row$tablename
-  
+
   if(!file.exists(paste0('tables/', tab, '.rda'))){
     r <- NULL
     attempt <- 1
@@ -55,8 +55,10 @@ by(allTables, 1:nrow(allTables), function(row){
 #Delete old samples of non-existent tables
 delete = data.frame(tablename=gsub('.rda', '', list.files('tables/', pattern="*.rda"))) %>% anti_join(allTables, by='tablename')
 
-res = by(delete, 1:nrow(delete), function(row){
-  tab = row$tablename
-  
-  file.remove(paste0('tables/',tab,'.rda'))
-})
+if(nrow(delete) > 0){
+  res = by(delete, 1:nrow(delete), function(row){
+    tab = row$tablename
+
+    file.remove(paste0('tables/',tab,'.rda'))
+  })
+}
